Extracted common attributes and helpers from BST and linked lists
