+++ client/src/visualiser-src/binary-search-tree-visualiser/animation-producer/BSTRotateAnimationProducer.ts
@@ -1,7 +1,7 @@
 import { Line } from '@svgdotjs/svg.js';
 import BSTAnimationProducer from './BSTAnimationProducer';
 import { Node } from '../util/typedefs';
-import { getPointerStartEndCoordinates } from '../util/util';
+import { getPointerStartEndCoordinates } from '../../common/helpers';
 import { rotateLeftCodeSnippet, rotateRightCodeSnippet } from '../util/codeSnippets';
 
 export default class BSTRotateAnimationProducer extends BSTAnimationProducer {
