+++ client/src/visualiser-src/binary-search-tree-visualiser/util/settings.ts
@@ -1,6 +1,4 @@
-// this file is used to store visual settings of the bst visualiser
-export const strokeWidth = 3;
-
+import { strokeWidth } from '../../common/constants';
 export const nodeStyle = {
   fill: '#ffffff',
   stroke: '#000',
@@ -25,5 +23,3 @@ export const lineStyle = {
 
 // we need this so that all svg elements have some padding
 export const canvasPadding = 75;
-export const markerLength = 15;
-export const pathD = `M 0 0 L ${markerLength} ${markerLength / 2} L 0 ${markerLength} z`;
