+++ client/src/visualiser-src/common/constants.ts
@@ -1,0 +1,6 @@
+export const strokeWidth = 3;
+export const nodeDiameter = 50;
+export const actualNodeDiameter = nodeDiameter + strokeWidth;
+
+export const markerLength = 15;
+export const pathD = `M 0 0 L ${markerLength} ${markerLength / 2} L 0 ${markerLength} z`;
